pyproject.toml
[tool.poetry]
name = "booking_service"
version = "0.1.0"
description = "A booking service for managing appointments"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-flask = "^1.3.0"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    "."
]